require version in release script

+++ release.sh
@@ -3,6 +3,10 @@
 set -e
 
 VERSION=$1
+if test -z "$VERSION"; then
+    echo "Give the version"
+    exit 1
+fi
 
 git tag $VERSION
 rm dist/*
